Switch govuk connect to use govuk-puppet for hosting detection
Rather than relying on what the govuk-developer-docs say regarding
where the Production version of the app is running.

Previously, this was impossible, as the hieradata wasn't an accurate
picture of what was running where, but now it's up to date, it's
possible to use it. This approach is also more accurate for Staging,
as currently some apps run in AWS in Staging, but Carrenza in
Production.