Fix the script to make it work again
Previously the script was run in a single namespace, with all methods
being defined at the top-level. This fixes the code in the new CLI
class so methods called from constants continue to work.

This also fixes a strange syntax issue with multiline help strings;
the syntax appears to be more strict when defined within a class.