Fix suppressing messages in test output
Previously a linting auto-correct changed most calls to STDERR.puts into
calls to "warn", which caused them to start leaking onto the console as
part of the test output. Calls to "warn" are part of core Ruby, and can't
be intercepted by the existing stub of STDERR.

This replaces the previous stub of STDERR with a native RSpec matcher
that has the effect of capturing the output. Although it's not ideal
to suppress warnings, since they may come from beyond our code, this
is easier than introducing a proxy class into the already-complicated CLI.

This also replaces a few remaining calls to STDERR.puts with a call to
the "print_empty_line" method, which now also calls "warn" with an empty
string (otherwise nothing is printed). I chose to rename the method
because it actually prints *two* newlines i.e. an empty line.