Downgrade min Ruby to 2.3
Previously we increased the minimum Ruby version to 2.4, and then 2.5,
noting that most people have 2.6 installed. However, when installing
via Homebrew, this causes an error since the formula calls out to system
Ruby, which for many is still 2.3 (OS Mojave).

This downgrades the minimum Ruby version back to 2.3, so that Homebrew
installations continue to work.