Remove staging carrenza from govuk-connect
The jumpbox in carrenza staging is no longer reachable, so there's no
need for it to be in govuk-connect anymore.

The situation at the moment is a little confusing, as if you don't
specify a hosting provider (e.g. `gds govuk connect -e staging
mirrorer`), it will attempt to get the list of nodes from carrenza
first. This doesn't work, because there's no jumpbox, instead it errors
in a way that doesn't clearly prompt the user to use the `aws/` prefix.

(The jumpbox is still there in production, along with a few other
machines, so that one can stay for now)