Automate pushing gem to RubyGems
Concourse currently just builds the gem and discards it.
This enables Concourse to also release the gem. The push
step is not idempotent - it will fail the job if the same
version is pushed twice.
Fix: https://github.com/troykinsella/concourse-rubygems-resource/pull/2

We'll still need to manually make a change to add new versions
to brew, but we can document this process:
https://github.com/alphagov/homebrew-gds/blob/master/Formula/govuk-connect.rb#L4