Fix dbconsole undefined local variable or method `name_and_number'
You might want to open a dbconsole on a specific numbered instance. For
example, so you can run a query and save the results to a file which you can
then SCP off the server.

At the moment there's a typo in the variable name, so it errors. Come back
static typing, all is forgiven.