Set a .ruby-version file
This removes the .ruby-version entry from .gitignore (it's a bit
annoying having this default to the system one) and sets the Ruby
version to the lowest one supported by the app. This allows us to drop
the ruby version from .rubocop.yml too.

This resolves a problem where ruby-version isn't known to run the
release process in CI.