Create a test_matrix GitHub Action job
This changes the configuration of the jobs to work around a problem with
the automated status checks. On GOV.UK repos we require that a status
check of "test" is met by a repo, however in matrix builds that are
named "test" there is no singular "test" check, instead there are
multiple "test (<distinguishing criteria>)" jobs and thus the status
check is never satisfied.

By having an individual test job that requires all the matrix jobs to
pass we can work around this issue. As per: https://github.community/t/status-check-for-a-matrix-jobs/127354/3